+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.53.0
 - Instrument [Winston](https://github.com/winstonjs/winston).
 - Exclude file system entries that are not directories in dependency analysis, fixes confusing warning.